Move merit list to top level of exercise
