+++ src/views/Exercise/MeritList.vue
